Updated the Project Flow
I separated the initialisation of the schedules and the method for changing what's outputted from each other.  This should allow me to update all of the schedules independently of each other.  I also set up initial output methods.  These will be used to make it easier to output to the command line.