Fixed a Bug with the SJF Schedule
There was an interesting bug where if a job arrived the second after a job finished the first job would run again.  This was because I was using an int to tell which job to run.  When the job was finished the int would increase by one.  In the next second a job would be added so that the size of the array would increase as well.  These two actions nullified each other so the first job would run twice.